chore(refactor): inlined module into funcs.go
