+++ pkg/cc/funcs.go
@@ -13,10 +13,10 @@ import (
 	"sort"
 	"strconv"
 	"strings"
+	"syscall"
 
 	"github.com/1898andCo/HAOS/pkg/command"
 	"github.com/1898andCo/HAOS/pkg/config"
-	"github.com/1898andCo/HAOS/pkg/hostname"
 	"github.com/1898andCo/HAOS/pkg/mode"
 	"github.com/1898andCo/HAOS/pkg/ssh"
 	"github.com/1898andCo/HAOS/pkg/system"
@@ -75,7 +75,46 @@ func ApplySysctls(cfg *config.CloudConfig) error {
 }
 
 func ApplyHostname(cfg *config.CloudConfig) error {
-	return hostname.SetHostname(cfg)
+	hostname := cfg.Hostname
+	if hostname == "" {
+		return nil
+	}
+	if err := syscall.Sethostname([]byte(hostname)); err != nil {
+		return err
+	}
+	return syncHostname()
+}
+
+func syncHostname() error {
+	hostname, err := os.Hostname()
+	if err != nil {
+		return err
+	}
+	if hostname == "" {
+		return nil
+	}
+
+	if err := afero.WriteFile(system.AppFs, "/etc/hostname", []byte(hostname+"\n"), 0644); err != nil {
+		return err
+	}
+
+	hosts, err := os.Open("/etc/hosts")
+	if err != nil {
+		return err
+	}
+	defer hosts.Close()
+	lines := bufio.NewScanner(hosts)
+	content := ""
+	for lines.Scan() {
+		line := strings.TrimSpace(lines.Text())
+		fields := strings.Fields(line)
+		if len(fields) > 0 && fields[0] == "127.0.1.1" {
+			content += "127.0.1.1 " + hostname + "\n"
+			continue
+		}
+		content += line + "\n"
+	}
+	return afero.WriteFile(system.AppFs, "/etc/hosts", []byte(content), 0600)
 }
 
 func ApplyPassword(cfg *config.CloudConfig) error {
